fix: bug fix about when fb and google user log in with firebase realtime
If users that fb and google, firebase will make child which name of "users".

The users is array type, it's got email and name.

And basically make at first time which is users sign in that users is not exsist.

But after fb and google user sign in, make another element at the firebase realtime database - the "users" array.

I found error code at DatabaseManager.swift, the function name of "userExist" is occured bug.

In that function had to type cast that snapShot value to [String: Any] dictionary.

But I did typecasted by "String".

So I found that code that occured bug, fixed to [String: Any].