bug: message UI bug.
When recevier user send message to sender user, the recevier user's message UI is not changed.

When the recevier send to the sender that message it have to change message UI's color and position.

If I sended message, the message UI color must be green and position is right side.

Otherwise if I receved message, the message UI color must be gray and position is left side.

But this bug is after first message, notting change the message UI.