fix: fixed video data type to url type, not searching myself.
1. fixed video data type.
When playing video message, It doesen't play because the video url is converted by value type which 'Data'.
So, fixed data type to the type of value that 'URL'

2. fixed search method.
When searching user which myself, It's not my intened result
When I searched my name, It has to find myself.
So, I fixed that bug, which added new condition code.
I used email and current user email filterd the user, If the user is myself that return false.
So, the search view can show the text that 'no result'.